Отредактировал согласно замечаний.Автопарк лвл 4
