+++ src/main/java/com/Autopark/Color.java
@@ -15,4 +15,3 @@ public enum Color {
         return name();
     }
 }
-
