+++ src/main/java/com/Autopark/Main.java
@@ -1,5 +1,9 @@
 package com.Autopark;
 
+import com.Autopark.Engine.DieselEngine;
+import com.Autopark.Engine.ElectricalEngine;
+import com.Autopark.Engine.GasolineEngine;
+
 public class Main {
 
     public static final int SIZE_VEHICLES = 7;
@@ -13,6 +17,7 @@ public class Main {
         print.printlnSortedVehicleType(initVehicle());
         print.printlnMaxAndMinMileage(initVehicle());
         print.printlnEqualVehicleType(initVehicle());
+        print.printAutoMaxKilometrs(initVehicle());
 
     }
 
@@ -56,19 +61,19 @@ public class Main {
         VehicleType[] as = initVehicleType();
         Vehicle[] vehicle = new Vehicle[7];
         vehicle[0] = new Vehicle(as[0], "Volkswagen Crafter", "5427 AX-7", 2022,
-                2015, 376000, Color.BLUE, 75);
+                2015, 376000, Color.BLUE, 75, new GasolineEngine(2, 8.1, 75));
         vehicle[1] = new Vehicle(as[0], "Volkswagen Crafter", "6427 AA-7", 2500,
-                2014, 227010, Color.WHITE, 75);
+                2014, 227010, Color.WHITE, 75, new GasolineEngine(2.18, 8.5, 75));
         vehicle[2] = new Vehicle(as[0], "ElectricBus", "6785 ВA-7", 12080,
-                2019, 20451, Color.GREEN, 100);
+                2019, 20451, Color.GREEN, 100, new ElectricalEngine(50, 150));
         vehicle[3] = new Vehicle(as[0], "Golf 5", "8682 AX-7", 1200,
-                2006, 230451, Color.GRAY, 55);
+                2006, 230451, Color.GRAY, 55, new DieselEngine(1.6, 7.2, 55));
         vehicle[4] = new Vehicle(as[0], "TestaModel S70D", "0001 AA-7", 2200,
-                2019, 10454, Color.WHITE, 70);
+                2019, 10454, Color.WHITE, 70, new ElectricalEngine(25, 70));
         vehicle[5] = new Vehicle(as[2], "Hamm HD", null, 3000,
-                2016, 122, Color.YELLOW, 135);
+                2016, 122, Color.YELLOW, 135, new DieselEngine(3.2, 25, 20));
         vehicle[6] = new Vehicle(as[3], "MT3", "1145 AB-7", 1200,
-                2020, 109, Color.RED, 130);
+                2020, 109, Color.RED, 130, new DieselEngine(4.75, 20.1, 135));
         return vehicle;
     }
 
@@ -102,7 +107,6 @@ public class Main {
         }
         return maxMileageVehicle;
     }
-
     public static class Helper {
         public static <T> void displayArr(T[] elements) {
             for (T element :
@@ -110,6 +114,22 @@ public class Main {
                 System.out.println(element.toString());
             }
             System.out.println();
+        }
+    }
+
+    public static void autoMaxKilometrs(Vehicle[] vehicle) {
+        double maxKilometers = 0.0;
+        for (Vehicle as : vehicle) {
+            if (as.getEngine().getMaxKilometers() - maxKilometers > 0) {
+                maxKilometers = as.getEngine().getMaxKilometers();
+            }
+        }
+
+        for (Vehicle as : vehicle) {
+            if (maxKilometers == as.getEngine().getMaxKilometers()) {
+                System.out.println(as);
+
+            }
         }
     }
 }
