--- src/main/java/level1and2/Color.java
@@ -1,17 +1,0 @@
-package level1and2;
-
-public enum Color {
-
-    RED,
-    WHITE,
-    BLUE,
-    GREY,
-    BLACK,
-    GREEN,
-    YELLOW,
-    GRAY;
-
-    public String getColor() {
-        return name();
-    }
-}
